sast/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
reportlab==4.0.0
matplotlib==3.8.0
python-multipart==0.0.9
qrcode[pil]==7.4.2
semgrep==1.52.0
bandit==1.7.7
python-dotenv==1.0.1
lxml==4.9.3
defusedxml==0.7.1
tabulate==0.9.0
rich==13.6.0

# Added for RDS and ElastiCache integration
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1

